+++ Changes.md
@@ -5,6 +5,7 @@ Fixes
 -----
 
 - LightPositionTool : Fixed bug where a light would rotate around its local Z-axis during placement.
+- RotateTool : Fixed bug where objects would rotate around their local Z-axis when using aim at target mode.
 
 1.5.15.0 (relative to 1.5.14.0)
 ========